Update the toolkit to 1.8.4
Update Dojo version to 1.8.4. It has been a long time the version
hasn't been updated! Thanks to Bamboo this commit make the gem up
to date with the new last 1.8.x release. We will wait for the first
patch before updating to 1.9.x or 2.0 which will be in 1.0 I hope.

Update the Dojo::VERSION constant as well and add this statement
to the contribution guideline to make sure users can know the current
toolkit version quickly!

:sparkles: